Refactor: GlobalExceptionHandler를 exception 패키지로 이동 및 기존 폴더 정리
