+++ Backend/src/main/java/com/backend/global/exception/GlobalExceptionHandler.java
@@ -2,51 +2,20 @@ package com.backend.global.exception;
 
 import com.backend.global.rsData.RsData;
 import org.springframework.http.ResponseEntity;
-import org.springframework.http.converter.HttpMessageNotReadableException;
-import org.springframework.validation.FieldError;
-import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.ResponseBody;
-
-import java.util.Comparator;
-import java.util.stream.Collectors;
 
 @ControllerAdvice
 public class GlobalExceptionHandler {
 
     @ExceptionHandler(ServiceException.class)
-    @ResponseBody
-    public ResponseEntity<RsData<Void>> handleException(ServiceException e) {
-        RsData<Void> body = e.getRsData();           // resultCode / msg 유지
-        int status = body.getStatusCode();           // "404-1" → 404
-        return ResponseEntity.status(status).body(body);
-    }
-
-    @ExceptionHandler(MethodArgumentNotValidException.class)
-    @ResponseBody
-    public RsData<Void> handleException(MethodArgumentNotValidException e) {
-        String message = e.getBindingResult()
-                .getAllErrors()
-                .stream()
-                .filter(error -> error instanceof FieldError)
-                .map(error -> (FieldError) error)
-                .map(error -> error.getField() + "-" + error.getCode() + "-" + error.getDefaultMessage())
-                .sorted(Comparator.comparing(String::toString))
-                .collect(Collectors.joining("\n"));
-
-        return new RsData<Void>(
-                "400-1",
-                message
-        );
-    }
-
-    @ExceptionHandler(HttpMessageNotReadableException.class)
-    @ResponseBody
-    public RsData<Void> handleException(HttpMessageNotReadableException e) {
-        return new RsData<Void>(
-                "400-2",
-                "잘못된 형식의 요청 데이터입니다."
-        );
+    public ResponseEntity<RsData<Void>> handleServiceException(ServiceException e) {
+        var errorCode = e.getErrorCode();
+        return ResponseEntity
+                .status(errorCode.getStatus())
+                .body(new RsData<>(
+                        String.valueOf(errorCode.getCode()),
+                        errorCode.getMessage()
+                ));
     }
 }
