--- Backend/src/main/java/com/backend/global/globalExceptionHandler/GlobalExceptionHandler.java
@@ -1,22 +1,0 @@
-package com.backend.global.globalExceptionHandler;
-
-import com.backend.global.exception.ServiceException;
-import com.backend.global.rsData.RsData;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-
-@ControllerAdvice
-public class GlobalExceptionHandler {
-
-    @ExceptionHandler(ServiceException.class)
-    public ResponseEntity<RsData<Void>> handleServiceException(ServiceException e) {
-        var errorCode = e.getErrorCode();
-        return ResponseEntity
-                .status(errorCode.getStatus())
-                .body(new RsData<>(
-                        String.valueOf(errorCode.getCode()),
-                        errorCode.getMessage()
-                ));
-    }
-}
